+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add `average` method
 - Add `arrMove` method
+- Add `arrTo1dArray` method
 
 ## 0.0.3 (2024-10-02)